backend/pyproject.toml
[project]
name = "get-publications"
version = "0.1.0"
description = "Grab publications from PubMed"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "biopython>=1.84",
  "polars>=1.5.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
  "ruff>=0.6.1",
  "mypy>=1.11.1",
  "pre-commit>=3.8.0"
]

[tool.mypy]
ignore_missing_imports = true
allow_redefinition = true
disable_error_code = ["no-redef"]

[tool.ruff]
line-length = 88
indent-width = 4
src = ["src"]
target-version = "py312"

[tool.ruff.lint]
select = ["ANN", "D", "E", "F", "I"]
ignore = [
  "ANN002",  # args should not be annotated.
  "ANN003",  # kwargs should not be annotated.
  "ANN101",  # self should not be annotated.
  "ANN102",  # cls should not be annotated.
  "D107"  # undocumented-public-init (e.g. __init__)
]
fixable = ["ALL"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
